requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0

# Data Validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Web Scraping
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
selenium==4.15.0

# Notion Integration
notion-client==2.2.1

# Task Scheduling
apscheduler==3.10.4

# Caching
redis==5.0.1

# Environment Variables
python-dotenv==1.0.0

# Logging and Monitoring
structlog==23.2.0
sentry-sdk==1.38.0

# Data Processing
pandas==2.1.4
numpy==1.24.3

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development Tools
black==23.11.0
flake8==6.1.0
mypy==1.7.1